introduces Notifier, to mediate between a Producer and a Subscriber
The Subscriber communicates with its Producer using the Subscription.
The Producer notifies each of its Subscribers using their Notifier.

The Subscription holds a strong reference to a Producer
The Notifier holds a weak reference to a Subscriber.
When a Subscriber is deinitialized, it cancels its Subscription. [note: this MUST happen]
The cancellation of a Subscription leads the Producer to release both the Subscription and the Notifier

The final() operator has been successfully translated to this new system.
The testLifetime() tests now work (and have been expanded)